+++ components/DesktopSidebar/UserIcon.tsx
